Extract default-depth-settings UI logic into engine/src/depth/defaultSettings.ts
Moves updateDefaultButtonState, updateSingleDefaultButtonState,
useAsDefaultSettings, resetToDefaultSettings, resetMonoParameters,
resetDisparityOffset, and resetPrinciplePoint out of
PointCloudVisualizer - the "star as default" / "reset to default"
depth-settings feature, including the fairly involved
current-vs-default comparison logic in
updateSingleDefaultButtonState. 20+ call sites across event listener
wiring kept working unchanged via thin delegating wrappers.

main.ts: 11,786 -> 11,547 lines.

Verified: engine tsc clean, full webpack compile clean, all 59
extension tests passing under xvfb-run.
